+++ src/constant/index.ts
@@ -246,11 +246,7 @@ export const ALL_THEME = [
   ...(window.ksu ? ['light-monet', 'dark-monet'] : []),
   'halloween',
   'forest',
-  'lofi',
-  'wireframe',
-  'black',
   'dracula',
-  'business',
   'night',
   'dim',
   'nord',
@@ -299,6 +295,7 @@ export const DEFAULT_THEME = {
 export type THEME = Record<string, string>
 
 export enum IP_INFO_API {
+  IPIP = 'ipip.net',
   IPSB = 'ip.sb',
   IPWHOIS = 'ipwho.is',
   IPAPI = 'ipapi.is',